requirements_food.txt
# 食物分类项目额外依赖
# Additional dependencies for food classification project

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0

# 数据处理
pandas>=1.3.0
numpy>=1.21.0

# 可视化（可选）
matplotlib>=3.3.0
seaborn>=0.11.0

# 进度条
tqdm>=4.62.0

# 日志记录（可选）
wandb>=0.12.0
tensorboard>=2.7.0

# 数据增强（可选）
albumentations>=1.1.0

# 模型导出（可选）
onnx>=1.10.0
onnxruntime>=1.9.0
